+++ parser/sql/statement/core/src/main/java/org/apache/shardingsphere/sql/parser/statement/core/statement/ddl/AlterTableStatement.java
@@ -93,6 +93,14 @@ public abstract class AlterTableStatement extends AbstractSQLStatement implement
     }
     
     /**
+     * Set modify collection retrieval.
+     *
+     * @param modifyCollectionRetrieval modify collection retrieval
+     */
+    public void setModifyCollectionRetrieval(final ModifyCollectionRetrievalSegment modifyCollectionRetrieval) {
+    }
+    
+    /**
      * Get modify collection retrieval.
      *
      * @return modify collection retrieval