+++ parser/sql/statement/type/presto/src/main/java/org/apache/shardingsphere/sql/parser/statement/presto/ddl/PrestoDropTableStatement.java
@@ -19,6 +19,7 @@ package org.apache.shardingsphere.sql.parser.statement.presto.ddl;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
+import lombok.Setter;
 import org.apache.shardingsphere.sql.parser.statement.core.statement.ddl.DropTableStatement;
 import org.apache.shardingsphere.sql.parser.statement.presto.PrestoStatement;
 
@@ -27,7 +28,8 @@ import org.apache.shardingsphere.sql.parser.statement.presto.PrestoStatement;
  */
 @RequiredArgsConstructor
 @Getter
+@Setter
 public final class PrestoDropTableStatement extends DropTableStatement implements PrestoStatement {
     
-    private final boolean ifExists;
+    private boolean ifExists;
 }